+++ cashflows/settings/__init__.py
@@ -1,0 +1,3 @@
+FLOW_TYPES = ["inflow", "outflow"]
+TIME_TYPES = ["int", "date"]
+DEFAULT_CURRENCY = "mxn"
